implemented test organization improvements from discussion
